% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatTables.R
\name{formatSequenceSymmetry}
\alias{formatSequenceSymmetry}
\title{A formatted visualization of sequence_symmetry objects.}
\usage{
formatSequenceSymmetry(
  result,
  type = "gt",
  estimateNameFormat = c(`N (\%)` = "<count> (<percentage> \%)", `SR (CI)` =
    "<point_estimate> (<lower_CI> - <upper_CI>)"),
  style = "default",
  crude = TRUE,
  adjusted = TRUE,
  studyPopulation = TRUE,
  indexName = TRUE,
  markerName = TRUE,
  cdmName = TRUE,
  .options = NULL
)
}
\arguments{
\item{result}{A sequence_symmetry object.}

\item{type}{Type of desired formatted table, possibilities: "gt",
"flextable", "tibble".}

\item{estimateNameFormat}{.}

\item{style}{.}

\item{.options}{decimals}

\item{header}{.}

\item{splitGroup}{.}
}
\value{
A table with a formatted version of the sequence_symmetry object.
}
\description{
It provides a formatted table with the contents of the getSequenceRatios
output.
}
\examples{
{
}


}
