% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatTables.R
\name{formatSequenceSymmetryOptions}
\alias{formatSequenceSymmetryOptions}
\title{A formatted visualization of sequence_symmetry objects.}
\usage{
formatSequenceSymmetryOptions()
}
\value{
The default .options named list.
}
\description{
It provides a list of allowed inputs for .option argument in
formatSequenceSymmetry and their given default value.
}
\examples{
{
}


}
