% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDrugCohortPSSA.R
\name{generateDrugCohortPSSA}
\alias{generateDrugCohortPSSA}
\title{Generate PSSA Drug Cohort}
\usage{
generateDrugCohortPSSA(
  cdm,
  index,
  marker,
  table_name = "pssa",
  prior_obs = 365,
  start_date,
  end_date
)
}
\arguments{
\item{cdm}{A CDM reference.}

\item{index}{Index drug(s), need to specify the name and the level in the ATC format.}

\item{marker}{Marker drug(s), need to specify the name and the level in the ATC format.}

\item{table_name}{Table name in CDM of user's choice.}

\item{prior_obs}{Prior observation that the user would like to impose on both index drug(s) and marker drug(s).}

\item{start_date}{Start date that the user would impose so that both index drug(s) and marker drug(s) would be initiated after this day. Set NA if this is not necessary.}

\item{end_date}{End date that the user would impose so that both index drug(s) and marker drug(s) would be initiated before this day. Set NA if this is not necessary.}
}
\description{
Generate the necessary drug cohorts based on the index drug(s) and
marker drug(s) specified. The output of this function will subsequently
be fed into, for exmaple, getPSSA().
}
