% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCohortSequence.R
\name{getCohortSequence}
\alias{getCohortSequence}
\title{Intersecting the index and marker cohorts prior to calculating Sequence Symmetry Ratios}
\usage{
getCohortSequence(
  cdm,
  indexTable,
  markerTable,
  name = "joined_cohorts",
  dateRange = as.Date(c(NA, NA)),
  indexId = NULL,
  markerId = NULL,
  daysPriorObservation = 0,
  washoutWindow = 0,
  indexMarkerGap = NULL,
  combinationWindow = c(0, 365)
)
}
\arguments{
\item{cdm}{A CDM reference.}

\item{indexTable}{A table in the CDM that the index cohorts should come from.}

\item{markerTable}{A table in the CDM that the marker cohorts should come from.}

\item{name}{the name within the cdm that the output is called. Default is joined_cohorts.}

\item{dateRange}{Two dates indicating study period and the sequences that the user wants
to restrict to.}

\item{indexId}{Cohort definition IDs in indexTable to be considered for the analysis.
Change to NULL if all indices are wished to be included.}

\item{markerId}{Cohort definition IDs in markerTable to be considered for the analysis.
Change to NULL if all markers are wished to be included.}

\item{daysPriorObservation}{The minimum amount of prior observation required on both the index
and marker cohorts per person.}

\item{washoutWindow}{A washout window to be applied on both the index cohort event and marker cohort.}

\item{indexMarkerGap}{The time before the start of the second episode
of the drug (could be either marker or index) and the time after the end of the first
episode (could be either marker or index).}

\item{combinationWindow}{a constrain to be placed on the gap between two iniations.
Default c(0,365), meaning the gap should be larger than 0 but less than or equal to 365.}
}
\value{
A table in the cdm reference.
}
\description{
Join two tables in the CDM (one for index and the other for marker cohorts)
into a new table in the cdm taking into account the maximum time interval between events.
Index and marker cohorts should be instantiated in advance by the user.
}
\examples{
\donttest{
library(PatientProfiles)
cdm <- mockPatientProfiles()
cdm <- CohortSymmetry::getCohortSequence(
  cdm = cdm,
  indexTable = "cohort1",
  markerTable = "cohort2"
)
## default name - joined_cohorts
 cdm$joined_cohorts
}
}
