% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHistogram.R
\name{getHistogram}
\alias{getHistogram}
\title{Histogram for sequences}
\usage{
getHistogram(pssa_output, time_scale = "weeks")
}
\arguments{
\item{pssa_output}{the output from PSSA functions (for example, getPSSA(), getPSSAStrata() or getPSSASubset())}

\item{time_scale}{the x axis of the histogram plot, options are weeks, days or months}
}
\description{
Histogram for sequences
}
