% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPSSAStrata.R
\name{getPSSAStrata}
\alias{getPSSAStrata}
\title{PSSA in age and sex strata}
\usage{
getPSSAStrata(
  cdm,
  ageGroup,
  sex,
  index,
  marker,
  prior_obs = 365,
  start_date = NA,
  end_date = NA,
  table_name = "pssa",
  study_time = NULL,
  confidence_interval_level = 0.025
)
}
\arguments{
\item{cdm}{A CDM reference.}

\item{ageGroup}{Age groups, for example: list(c(0,17), c(18,30))}

\item{sex}{Sex groups, for example: list(c("Male", "Female", "Both"))}

\item{index}{Index drug(s), need to specify the name and the level in the ATC format.}

\item{marker}{Marker drug(s), need to specify the name and the level in the ATC format.}

\item{prior_obs}{Prior observation that the user would like to impose on both index drug(s) and marker drug(s).}

\item{start_date}{Start date that the user would impose so that both index drug(s) and marker drug(s) would be initiated after this day. Set NA if this is not necessary.}

\item{end_date}{End date that the user would impose so that both index drug(s) and marker drug(s) would be initiated before this day. Set NA if this is not necessary.}

\item{table_name}{Table name in CDM of user's choice.}

\item{study_time}{Study time is defined to be the gap between the initiation of index and the initiation of marker. Default is NULL, meaning no restriction is imposed.}

\item{confidence_interval_level}{Default 0.025, meaning 95\% interval. If the user wants N\% interval (N between 0 and 100) then change the default to (1-N/100)/2.}
}
\description{
PSSA calculations with different age and sex strata
}
