% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPSSASubset.R
\name{getPSSASubset}
\alias{getPSSASubset}
\title{PSSA calculations with the underlying population restricted by a predefined JSON file.}
\usage{
getPSSASubset(
  cdm,
  index,
  marker,
  subset_name,
  subset_id,
  table_name = "pssa",
  study_time = NULL,
  confidence_interval_level = 0.025
)
}
\arguments{
\item{cdm}{A CDM reference.}

\item{index}{Index drug(s), need to specify the name and the level in the ATC format.}

\item{marker}{Marker drug(s), need to specify the name and the level in the ATC format.}

\item{subset_name}{A table in the CDM that the user wants to subset the underlying by.}

\item{subset_id}{The cohort definition ID of the denominator cohort of interest.}

\item{table_name}{Table name in CDM of user's choice.}

\item{study_time}{Study time is defined to be the gap between the initiation of index and the initiation of marker. Default is NULL, meaning no restriction is imposed.}

\item{confidence_interval_level}{Default 0.025, meaning 95\% interval. If the user wants N\% interval (N between 0 and 100) then change the default to (1-N/100)/2.}
}
\description{
PSSA calculations with the underlying population restricted by a predefined JSON file.
}
