% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSequenceRatios.R
\name{getSequenceRatios}
\alias{getSequenceRatios}
\title{Sequence ratio calculations}
\usage{
getSequenceRatios(cdm, outcomeTable, confidenceIntervalLevel = 0.025)
}
\arguments{
\item{cdm}{A CDM reference.}

\item{outcomeTable}{A table in the CDM that the output of getCohortSequence resides.}

\item{confidenceIntervalLevel}{Default is 0.025, which is the central 95\% confidence interval.}
}
\value{
A local table with all analyses with
}
\description{
Using getCohortSequence to obtain sequence ratios for the desired outcomes.
}
