% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWaitingTimeDistribution.R
\name{getWaitingTimeDistribution}
\alias{getWaitingTimeDistribution}
\title{Waiting Time Distribution}
\usage{
getWaitingTimeDistribution(
  cdm,
  drug,
  table_name = "wtd",
  start_date,
  end_date,
  prior_obs = 365
)
}
\arguments{
\item{cdm}{A CDM reference.}

\item{drug}{The drug(s) that the user wishes to compute the Waiting Time Distribution on.}

\item{table_name}{Table name in CDM of user's choice.}

\item{start_date}{Start date that the user would impose so that both index drug(s) and marker drug(s) would be initiated after this day. Set NA if this is not necessary.}

\item{end_date}{End date that the user would impose so that both index drug(s) and marker drug(s) would be initiated before this day. Set NA if this is not necessary.}

\item{prior_obs}{Prior observation that the user would like to impose on both index drug(s) and marker drug(s).}
}
\description{
Classically, waiting time distribution is required to compute null sequence ratio.
Though this is not what has been done in this package,
Waiting Time Distribution is included for competition.
For more information, see Preiss AK, Roughead EE, Pratt NL. Sequence symmetry analysis graphic adjustment for prescribing trends. BMC Medical Research Methodology. 2019 Jul 9;19(1).
}
