% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseSequenceRatio.R
\name{summariseSequenceRatio}
\alias{summariseSequenceRatio}
\title{Sequence ratio calculations}
\usage{
summariseSequenceRatio(
  cohort,
  cohortId = NULL,
  confidenceInterval = 95,
  movingAverageRestriction = 548
)
}
\arguments{
\item{cohort}{A cohort table in the cdm.}

\item{cohortId}{The Ids in the cohort that are to be included in the analyses.}

\item{confidenceInterval}{Default is 95, indicating the central 95\% confidence interval.}

\item{movingAverageRestriction}{The moving window when calculating nSR, default is 548.}
}
\value{
A local table with all the analyses.
}
\description{
Using generateSequenceCohortSet to obtain sequence ratios for the desired outcomes.
}
\examples{
\donttest{
library(CohortSymmetry)
cdm <- mockCohortSymmetry()
cdm <- generateSequenceCohortSet(cdm = cdm,
                                 name = "joined_cohorts",
                                 indexTable = "cohort_1",
                                 markerTable = "cohort_2")
pssa_result <- summariseSequenceRatio(cohort = cdm$joined_cohorts)
pssa_result
CDMConnector::cdmDisconnect(cdm)
}

}
