% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayTables.R
\name{tableSequenceRatiosOptions}
\alias{tableSequenceRatiosOptions}
\title{A formatted visualization of sequence_ratios objects.}
\usage{
tableSequenceRatiosOptions()
}
\value{
The default .options named list.
}
\description{
It provides a list of allowed inputs for .option argument in
tableSequenceRatios and their given default value.
}
\examples{
{
 tableSequenceRatiosOptions()
}

}
