% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidySequenceRatio.R
\name{tidySequenceSymmetry}
\alias{tidySequenceSymmetry}
\title{A tidy visualization of sequence_symmetry objects.}
\usage{
tidySequenceSymmetry(result, minCellCount)
}
\arguments{
\item{result}{A sequence_symmetry object.}

\item{minCellCount}{Minimum number of counts to report.}
}
\value{
A tibble with a tidy version of the sequence_symmetry
object.
}
\description{
It provides a tidy dataframe with the contents of the getSequenceRatios
output.
}
\examples{
{
}

}
