% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSSA-package.R
\docType{package}
\name{CohortSymmetry-package}
\alias{CohortSymmetry}
\alias{CohortSymmetry-package}
\title{CohortSymmetry: Sequence Symmetry Analysis using OMOP CDM (Common Data Model)}
\description{
Calculating crude sequence ratio (csr), adjusted sequence ratio (asr) and confidence intervals using data mapped to the Observational Medical Outcomes Partnership (OMOP) common data model. Calculations can also be done in a particular disease strata and stratification cohorts.
}
\author{
\strong{Maintainer}: Xihang Chen \email{xihang.chen@ndorms.ox.ac.uk} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
