% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatTables.R
\name{formatSequenceSymmetry}
\alias{formatSequenceSymmetry}
\title{A formatted visualization of sequence_symmetry objects.}
\usage{
formatSequenceSymmetry(
  result,
  type = "gt",
  estimateNameFormat = c(`N (\%)` = "<count> (<percentage> \%)", `SR (CI)` =
    "<point_estimate> (<lower_CI> - <upper_CI>)"),
  style = "default",
  crude = TRUE,
  adjusted = TRUE,
  studyPopulation = TRUE,
  indexName = TRUE,
  markerName = TRUE,
  cdmName = TRUE,
  .options = NULL
)
}
\arguments{
\item{result}{A sequence_symmetry object.}

\item{type}{Type of desired formatted table, possibilities: "gt",
"flextable", "tibble".}

\item{estimateNameFormat}{.}

\item{style}{Named list that specifies how to style the different parts of a
gt table or flextable. See visOmopResults package for more information on
how to define a style. Alternatively, use "default" to get visOmopResults
style, or NULL for gt/flextable default styling.}

\item{crude}{whether to report crude results.}

\item{adjusted}{whether to report adjusted results.}

\item{studyPopulation}{whether to report the study population.}

\item{indexName}{whether to report index names.}

\item{markerName}{whether to report marker names}

\item{cdmName}{whether to report database names.}

\item{.options}{named list with additional formatting options.
CohortSymmetry::formatSequenceSymmetryOptions() shows allowed arguments and
their default values.}
}
\value{
A formatted version of the sequence_symmetry object.
}
\description{
It provides a formatted table with the contents of the getSequenceRatios
output.
}
