% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDrugCohortPSSA.R
\name{generateDrugCohortPSSA}
\alias{generateDrugCohortPSSA}
\title{Title}
\usage{
generateDrugCohortPSSA(
  cdm,
  index,
  marker,
  table_name = "pssa",
  prior_obs = 365,
  start_date,
  end_date
)
}
\arguments{
\item{cdm}{a CDM reference}

\item{index}{index drug}

\item{marker}{marker drug}

\item{table_name}{table name in CDM}

\item{prior_obs}{prior observation}

\item{start_date}{start date}

\item{end_date}{end date}
}
\description{
Title
}
