% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCohortSequence.R
\name{getCohortSequence}
\alias{getCohortSequence}
\title{Intersecting the index and marker cohorts prior to calculating Sequence Symmetry Ratios}
\usage{
getCohortSequence(
  cdm,
  indexTable,
  markerTable,
  name = "joined_cohorts",
  dateRange = as.Date(c(NA, NA)),
  indexId = NULL,
  combineIndex = NULL,
  markerId = NULL,
  combineMarker = NULL,
  daysPriorObservation = 0,
  indexWashout = 0,
  markerWashout = 0,
  continuedExposureInterval = NULL,
  blackOutPeriod = 0,
  timeGap = 365
)
}
\arguments{
\item{cdm}{A CDM reference.}

\item{indexTable}{A table in the CDM that the index cohorts should come from.}

\item{markerTable}{A table in the CDM that the marker cohorts should come from.}

\item{name}{the name within the cdm that the output is called. Default is joined_cohorts.}

\item{dateRange}{Two dates indicating study period and the sequences that the user wants
to restrict to.}

\item{indexId}{Cohort definition IDs in indexTable to be considered for the analysis.
Change to NULL if all indices are wished to be included.}

\item{combineIndex}{A parameter to combine certain cohorts of the indexTable together.
Default is Null, meaning every cohort_definition_id is considered separately.}

\item{markerId}{Cohort definition IDs in markerTable to be considered for the analysis.
Change to NULL if all markers are wished to be included.}

\item{combineMarker}{A parameter to combine certain cohorts of the markerTable together.
Default is Null, meaning every cohort_definition_id is considered separately.}

\item{daysPriorObservation}{The minimum amount of prior observation required on both the index
and marker cohorts per person.}

\item{indexWashout}{Washout period to be applied to the index cohort event.}

\item{markerWashout}{Washout period to be applied to the marker cohort event.}

\item{continuedExposureInterval}{The time before the start of the second episode
of the drug (could be either marker or index) and the time after the end of the first
episode (could be either marker or index).}

\item{blackOutPeriod}{The minimum time the ADR is expected to take place.
Default is 0, meaning excluding the cases that see both dates on the same day.}

\item{timeGap}{The time between two initiations of index and marker. Default is 365.
Change to Inf if no constrains are imposed.}
}
\value{
A table in the cdm reference with subject_id, index_id, marker_id, index_date, marker_date, first_date and cdm_name.
}
\description{
Join two tables in the CDM (one for index and the other for marker cohorts)
into a new table in the cdm taking into account the maximum time interval between events.
Index and marker cohorts should be instantiated in advance by the user.
}
\examples{
\donttest{
library(PatientProfiles)
cdm <- mockPatientProfiles()
cdm <- CohortSymmetry::getCohortSequence(cdm = cdm,
                                         indexTable = "cohort1",
                                         markerTable = "cohort2")
## default name - joined_cohorts
cdm$joined_cohorts
}
}
