% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCohortSequence.R
\name{getCohortSequence}
\alias{getCohortSequence}
\title{Intersecting the index and marker cohorts prior to calculating Sequence Symmetry Ratios}
\usage{
getCohortSequence(
  cdm,
  indexTable,
  indexId = NULL,
  markerTable,
  markerId = NULL,
  timeGap = 365
)
}
\arguments{
\item{cdm}{A CDM reference.}

\item{indexTable}{A table in the CDM that the index cohorts should come from.}

\item{indexId}{Cohort definition IDs in indexTable to be considered for the analysis.
Change to NULL if all indices are wished to be included.}

\item{markerTable}{A table in the CDM that the marker cohorts should come from.}

\item{markerId}{Cohort definition IDs in markerTable to be considered for the analysis.
Change to NULL if all markers are wished to be included.}

\item{timeGap}{The time between two initiations of index and marker. Default is 365.
Change to Inf if no constrains are imposed.}
}
\value{
A table in the cdm reference with subject_id, index_id, marker_id, index_date, marker_date, first_date, time_gap and cdm_name.
}
\description{
Join two tables in the CDM (one for index and the other for marker cohorts)
into a new table in the cdm taking into account the maximum time interval between events.
Index and marker cohorts should be instantiated in advance
by the user and should only contain first-ever events.
}
