% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPSSA.R
\name{getPSSA}
\alias{getPSSA}
\title{Get PPSA}
\usage{
getPSSA(
  cdm,
  index,
  marker,
  cohort_table = NULL,
  table_name = "pssa",
  study_time = NULL,
  confidence_interval_level = 0.025,
  prior_obs = 365,
  start_date,
  end_date
)
}
\arguments{
\item{cdm}{A cdm reference}

\item{index}{Index drug}

\item{marker}{Marker drug}

\item{cohort_table}{Cohort table}

\item{table_name}{Table name}

\item{study_time}{Study time}

\item{confidence_interval_level}{Confidence interval level}

\item{prior_obs}{Prior observation}

\item{start_date}{Start date}

\item{end_date}{End date}
}
\description{
Get PPSA
}
