% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockCohortSymmetry.R
\name{mockCohortSymmetry}
\alias{mockCohortSymmetry}
\title{Creates mock cdm object for testing}
\usage{
mockCohortSymmetry(
  indexCohort,
  markerCohort,
  con = DBI::dbConnect(duckdb::duckdb(), ":memory:"),
  schema = "main"
)
}
\arguments{
\item{indexCohort}{the tibble of your index cohort}

\item{markerCohort}{the tibble of your marker cohort}

\item{con}{connection detail}

\item{schema}{name of your write schema}
}
\value{
mock cdm object contains your index and marker cohort
}
\description{
Creates mock cdm object for testing
}
\examples{
\donttest{
library(CohortSymmetry)
}
}
