% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSequenceRatio.R
\name{plotSequenceRatio}
\alias{plotSequenceRatio}
\title{A plot for the sequence ratios.}
\usage{
plotSequenceRatio(
  sequenceRatio,
  onlyaSR = FALSE,
  plotTitle = NULL,
  labs = c("SR", "Drug Pairs"),
  colours = c("red", "blue")
)
}
\arguments{
\item{sequenceRatio}{A table of the form of the output of summariseSequenceRatio.}

\item{onlyaSR}{If the only SR to be plotted is the adjusted SR.}

\item{plotTitle}{Title of the plot, if NULL no title will be plotted.}

\item{labs}{Axis labels for the plot.}

\item{colours}{Colours for both parts of the plot, pre- and post- time 0.}
}
\value{
A plot for the sequence ratios of index and marker cohorts.
}
\description{
It provides a ggplot of the sequence ratios of index and marker cohorts.
}
\examples{
\donttest{
library(CohortSymmetry)

cdm <- mockCohortSymmetry()
cdm <- generateSequenceCohortSet(cdm = cdm,
                                 indexTable = "cohort_1",
                                 markerTable = "cohort_2",
                                 name = "joined_cohort")
sr <- summariseSequenceRatio(cohort = cdm$joined_cohort)
plotSequenceRatio(sequenceRatio = sr)
CDMConnector::cdmDisconnect(cdm = cdm)
}
}
