% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTemporalSymmetry.R
\name{plotTemporalSymmetry}
\alias{plotTemporalSymmetry}
\title{A plot for the temporal symmetry of cohorts.}
\usage{
plotTemporalSymmetry(
  cdm,
  joinedTable,
  censorRange = c(1:4),
  xlim = c(-12, 12),
  colours = c("blue", "red")
)
}
\arguments{
\item{cdm}{A cdm object.}

\item{joinedTable}{The name of a table in the cdm of the form of the output
of generateSequenceCohortSet.}

\item{censorRange}{Counts to be censored, can be NULL.}

\item{xlim}{Limits for the x axis of the plot.}

\item{colours}{Colours for both parts of the plot, pre- and post- time 0.}
}
\value{
A plot for the temporal symmetry of cohorts.
}
\description{
It provides a ggplot of the temporal symmetry of two or more cohorts.
}
\examples{
\donttest{
library(CohortSymmetry)
cdm <- CohortSymmetry::mockCohortSymmetry()
cdm <- CohortSymmetry::generateSequenceCohortSet(cdm = cdm,
                                                 indexTable = "cohort_1",
                                                 markerTable = "cohort_2",
                                                 name = "joined_cohort")
plotTemporalSymmetry(cdm, "joined_cohort", censorRange = NULL)
CDMConnector::cdmDisconnect(cdm = cdm)
}
}
