% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTemporalSymmetry.R
\name{plotTemporalSymmetry}
\alias{plotTemporalSymmetry}
\title{A plot for the temporal symmetry of cohorts.}
\usage{
plotTemporalSymmetry(
  cdm,
  joinedTable,
  index_ids = NULL,
  marker_ids = NULL,
  plotTitle = NULL,
  labs = c("Time (months)", "Individuals (N)"),
  timescale = "month",
  censorRange = NULL,
  xlim = c(-12, 12),
  colours = c("blue", "red")
)
}
\arguments{
\item{cdm}{A cdm object.}

\item{joinedTable}{The name of a table in the cdm of the form of the output
of generateSequenceCohortSet.}

\item{index_ids}{What index ids to plot, if NULL all will be plotted.}

\item{marker_ids}{What marker ids to plot, if NULL all will be plotted.}

\item{plotTitle}{Title of the plot, if NULL no title will be plotted.}

\item{labs}{Axis labels for the plot.}

\item{timescale}{Timescale for the x axis of the plot (month, day, year).}

\item{censorRange}{Counts to be censored, can be NULL.}

\item{xlim}{Limits for the x axis of the plot.}

\item{colours}{Colours for both parts of the plot, pre- and post- time 0.}
}
\value{
A plot for the temporal symmetry of cohorts.
}
\description{
It provides a ggplot of the temporal symmetry of two or more cohorts.
}
\examples{
\donttest{
library(CohortSymmetry)
cdm <- CohortSymmetry::mockCohortSymmetry()
cdm <- CohortSymmetry::generateSequenceCohortSet(cdm = cdm,
                                                 indexTable = "cohort_1",
                                                 markerTable = "cohort_2",
                                                 name = "joined_cohort")
plotTemporalSymmetry(cdm, "joined_cohort", censorRange = NULL)
CDMConnector::cdmDisconnect(cdm = cdm)
}
}
