% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTemporalSymmetry.R
\name{plotTemporalSymmetry}
\alias{plotTemporalSymmetry}
\title{A plot for the temporal symmetry of cohorts.}
\usage{
plotTemporalSymmetry(
  cdm,
  sequenceTable,
  indexId = NULL,
  markerId = NULL,
  plotTitle = NULL,
  labs = c("Time (months)", "Individuals (N)"),
  timescale = "month",
  xlim = c(-12, 12),
  colours = c("blue", "red"),
  scales = "free"
)
}
\arguments{
\item{cdm}{A cdm object.}

\item{sequenceTable}{The name of a table in the cdm of the form of the output
of generateSequenceCohortSet.}

\item{indexId}{What index ids to plot, if NULL all will be plotted.}

\item{markerId}{What marker ids to plot, if NULL all will be plotted.}

\item{plotTitle}{Title of the plot, if NULL no title will be plotted.}

\item{labs}{Axis labels for the plot.}

\item{timescale}{Timescale for the x axis of the plot (month, day, year).}

\item{xlim}{Limits for the x axis of the plot.}

\item{colours}{Colours for both parts of the plot, pre- and post- time 0.}

\item{scales}{Whether to set free y scales for the facet wrap when there are
multiple plots (i.e. each plot has its own scaled y axis) or set them equal
for all. Only accepts "free" for the former and "fixed" for the latter.}
}
\value{
A plot for the temporal symmetry of cohorts.
}
\description{
It provides a ggplot of the temporal symmetry of two or more cohorts.
}
\examples{
\donttest{
library(CohortSymmetry)
cdm <- mockCohortSymmetry()
cdm <- generateSequenceCohortSet(cdm = cdm,
                                 indexTable = "cohort_1",
                                 markerTable = "cohort_2",
                                 name = "joined_cohort")
plotTemporalSymmetry(cdm = cdm,
                     sequenceTable = "joined_cohort")
CDMConnector::cdmDisconnect(cdm = cdm)
}
}
