% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseSequenceRatio.R
\name{summariseSequenceRatio}
\alias{summariseSequenceRatio}
\title{Sequence ratio calculations}
\usage{
summariseSequenceRatio(
  cdm,
  sequenceTable,
  sequenceId = NULL,
  confidenceInterval = 95,
  movingAverageRestriction = 548
)
}
\arguments{
\item{cdm}{A CDM reference.}

\item{sequenceTable}{A table in the CDM that the output of generateSequenceCohortSet resides.}

\item{sequenceId}{The Ids in the sequenceTable that are wished to be included in the analyses.}

\item{confidenceInterval}{Default is 95, indicating the central 95\% confidence interval.}

\item{movingAverageRestriction}{The moving window when calculating nSR, default is 548.}
}
\value{
A local table with all the analyses.
}
\description{
Using generateSequenceCohortSet to obtain sequence ratios for the desired outcomes.
}
\examples{
\donttest{
library(CohortSymmetry)
cdm <- CohortSymmetry::mockCohortSymmetry()
cdm <- CohortSymmetry::generateSequenceCohortSet(cdm = cdm,
                                                 name = "joined_cohorts",
                                                 indexTable = "cohort_1",
                                                 markerTable = "cohort_2")
pssa_result <- CohortSymmetry::summariseSequenceRatio (cdm = cdm,
                                                       sequenceTable = "joined_cohorts")
 pssa_result
 CDMConnector::cdmDisconnect(cdm)
}

}
