% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyTables.R
\name{tidySequenceSymmetry}
\alias{tidySequenceSymmetry}
\title{A tidy visualization of sequence_symmetry objects.}
\usage{
tidySequenceSymmetry(result)
}
\arguments{
\item{result}{A sequence_symmetry object.}
}
\value{
A tibble with a tidy version of the sequence_symmetry
object.
}
\description{
It provides a tidy dataframe with the contents of the summariseSequenceRatio
output.
}
\examples{
\donttest{
library(CohortSymmetry)
cdm <- mockCohortSymmetry()
cdm <- CohortSymmetry::generateSequenceCohortSet(cdm = cdm,
                                                 name = "joined_cohorts",
                                                 indexTable = "cohort_1",
                                                 markerTable = "cohort_2",
                                                 combinationWindow = c(0, Inf))
res <- CohortSymmetry::summariseSequenceRatio(
cdm = cdm,
sequenceCohortSet = "joined_cohorts")

tidy_result <- CohortSymmetry::tidySequenceSymmetry(res)

tidy_result
CDMConnector::cdmDisconnect(cdm = cdm)
}
}
